refactor: use tec tickets api for rsvp crud
