+++ src/Tickets/REST/TEC/V1/Traits/With_Filtered_Ticket_Params.php
@@ -131,24 +131,25 @@ trait With_Filtered_Ticket_Params {
 		}
 
 		$new_params = [
-			'id'                      => (int) ( $params['id'] ?? null ),
-			'event'                   => $event_id,
-			'ticket_id'               => $params['id'] ?? null,
-			'ticket_name'             => $params['title'] ?? $ticket_post->post_title ?? null,
-			'ticket_description'      => $params['content'] ?? $params['excerpt'] ?? $ticket_post->post_excerpt ?? null,
-			'ticket_price'            => $params['price'] ?? $ticket_data[ $orm->get_update_fields_aliases()['price'] ]['0'] ?? null,
-			'ticket_show_description' => $params['show_description'] ?? $ticket_data[ $orm->get_update_fields_aliases()['show_description'] ]['0'] ?? null,
-			'ticket_type'             => $params['type'] ?? $ticket_data[ $orm->get_update_fields_aliases()['type'] ]['0'] ?? null,
-			'ticket_sku'              => $params['sku'] ?? $ticket_data[ $orm->get_update_fields_aliases()['sku'] ]['0'] ?? null,
-			'ticket_start_date'       => $params['ticket_start_date'] ?? $ticket_data[ $orm->get_update_fields_aliases()['start_date'] ] ?? null,
-			'ticket_start_time'       => $params['ticket_start_time'] ?? $ticket_data[ $orm->get_update_fields_aliases()['start_time'] ] ?? null,
-			'ticket_end_date'         => $params['ticket_end_date'] ?? $ticket_data[ $orm->get_update_fields_aliases()['end_date'] ] ?? null,
-			'ticket_end_time'         => $params['ticket_end_time'] ?? $ticket_data[ $orm->get_update_fields_aliases()['end_time'] ] ?? null,
-			'tribe-ticket'            => array_filter( $tribe_ticket, fn( $value ) => null !== $value ),
-			'ticket_add_sale_price'   => isset( $params['sale_price'] ) || ! empty( $ticket_data[ $orm->get_update_fields_aliases()['sale_price_enabled'] ]['0'] ),
-			'ticket_sale_price'       => $params['sale_price'] ?? $ticket_data[ $orm->get_update_fields_aliases()['sale_price'] ]['0'] ?? null,
-			'ticket_sale_start_date'  => $params['sale_price_start_date'] ?? $ticket_data[ $orm->get_update_fields_aliases()['sale_price_start_date'] ]['0'] ?? null,
-			'ticket_sale_end_date'    => $params['sale_price_end_date'] ?? $ticket_data[ $orm->get_update_fields_aliases()['sale_price_end_date'] ]['0'] ?? null,
+			'id'                                => (int) ( $params['id'] ?? null ),
+			'event'                             => $event_id,
+			'ticket_id'                         => $params['id'] ?? null,
+			'ticket_name'                       => $params['title'] ?? $ticket_post->post_title ?? null,
+			'ticket_description'                => $params['content'] ?? $params['excerpt'] ?? $ticket_post->post_excerpt ?? null,
+			'ticket_price'                      => $params['price'] ?? $ticket_data[ $orm->get_update_fields_aliases()['price'] ]['0'] ?? null,
+			'ticket_show_description'           => $params['show_description'] ?? $ticket_data[ $orm->get_update_fields_aliases()['show_description'] ]['0'] ?? null,
+			'ticket_type'                       => $params['type'] ?? $ticket_data[ $orm->get_update_fields_aliases()['type'] ]['0'] ?? null,
+			'ticket_sku'                        => $params['sku'] ?? $ticket_data[ $orm->get_update_fields_aliases()['sku'] ]['0'] ?? null,
+			'ticket_start_date'                 => $params['ticket_start_date'] ?? $ticket_data[ $orm->get_update_fields_aliases()['start_date'] ] ?? null,
+			'ticket_start_time'                 => $params['ticket_start_time'] ?? $ticket_data[ $orm->get_update_fields_aliases()['start_time'] ] ?? null,
+			'ticket_end_date'                   => $params['ticket_end_date'] ?? $ticket_data[ $orm->get_update_fields_aliases()['end_date'] ] ?? null,
+			'ticket_end_time'                   => $params['ticket_end_time'] ?? $ticket_data[ $orm->get_update_fields_aliases()['end_time'] ] ?? null,
+			'tribe-ticket'                      => array_filter( $tribe_ticket, fn( $value ) => null !== $value ),
+			'ticket_add_sale_price'             => isset( $params['sale_price'] ) || ! empty( $ticket_data[ $orm->get_update_fields_aliases()['sale_price_enabled'] ]['0'] ),
+			'ticket_sale_price'                 => $params['sale_price'] ?? $ticket_data[ $orm->get_update_fields_aliases()['sale_price'] ]['0'] ?? null,
+			'ticket_sale_start_date'            => $params['sale_price_start_date'] ?? $ticket_data[ $orm->get_update_fields_aliases()['sale_price_start_date'] ]['0'] ?? null,
+			'ticket_sale_end_date'              => $params['sale_price_end_date'] ?? $ticket_data[ $orm->get_update_fields_aliases()['sale_price_end_date'] ]['0'] ?? null,
+			'ticket_rsvp_enable_cannot_go' => $params['show_not_going'] ?? null,
 		];
 
 		$new_params['ticket_sale_price']      = maybe_unserialize( $new_params['ticket_sale_price'] );
